Begin reverting acronym casing change

+++ sources/SDL/SDL/SDL3/Finger.gen.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // Ported from SDL.h and corresponding dependencies of SDL3.
 // Original source is Copyright (C) 1997-2024 Sam Lantinga. Licensed under the zlib license.
+using System;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
@@ -11,7 +12,7 @@ namespace Silk.NET.SDL;
 public partial struct Finger
 {
     [NativeName("id")]
-    public ulong Id;
+    public nuint Id;
 
     [NativeName("x")]
     public float X;
