+++ sources/SDL/SDL/SDL3/InitState.gen.cs
@@ -3,7 +3,6 @@
 // Ported from SDL.h and corresponding dependencies of SDL3.
 // Original source is Copyright (C) 1997-2024 Sam Lantinga. Licensed under the zlib license.
 using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
 
 namespace Silk.NET.SDL;
 
@@ -14,7 +13,7 @@ public unsafe partial struct InitState
     public AtomicInt Status;
 
     [NativeName("thread")]
-    public ulong Thread;
+    public nuint Thread;
 
     [NativeName("reserved")]
     public void* Reserved;
