+++ sources/SDL/SDL/SDL3/JoyBallEvent.gen.cs
@@ -3,7 +3,6 @@
 // Ported from SDL.h and corresponding dependencies of SDL3.
 // Original source is Copyright (C) 1997-2024 Sam Lantinga. Licensed under the zlib license.
 using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
 
 namespace Silk.NET.SDL;
 
@@ -17,7 +16,7 @@ public partial struct JoyBallEvent
     public uint Reserved;
 
     [NativeName("timestamp")]
-    public ulong Timestamp;
+    public nuint Timestamp;
 
     [NativeName("which")]
     public uint Which;
