+++ sources/SDL/SDL/SDL3/PathInfo.gen.cs
@@ -3,7 +3,6 @@
 // Ported from SDL.h and corresponding dependencies of SDL3.
 // Original source is Copyright (C) 1997-2024 Sam Lantinga. Licensed under the zlib license.
 using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
 
 namespace Silk.NET.SDL;
 
@@ -14,14 +13,14 @@ public partial struct PathInfo
     public PathType Type;
 
     [NativeName("size")]
-    public ulong Size;
+    public nuint Size;
 
     [NativeName("create_time")]
-    public long CreateTime;
+    public nint CreateTime;
 
     [NativeName("modify_time")]
-    public long ModifyTime;
+    public nint ModifyTime;
 
     [NativeName("access_time")]
-    public long AccessTime;
+    public nint AccessTime;
 }
